src/packages/win-computer-use/requirements.txt
# Core LLM clients (same as open-computer-use)
openai>=1.0.0
anthropic>=0.30.0

# Windows desktop control
pyautogui>=0.9.54
pyperclip>=1.9.0

# Fast screenshot capture
mss>=9.0.0

# Image processing
Pillow>=10.0.0

# Grounding models (HuggingFace Spaces)
gradio-client>=1.0.0

# Utilities
python-dotenv>=1.0.0
